Placed routines for small sorting networks into sort.h.
